pyproject.toml
[project]
name = "ipredict"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "attrs (==23.2.0)",
    "certifi (==2024.2.2)",
    "charset-normalizer (==3.3.2)",
    "click (==8.1.7)",
    "clickclick (==20.10.2)",
    "cmdstanpy (==1.2.1)",
    "connexion (==2.14.2)",
    "contourpy (==1.2.0)",
    "cycler (==0.12.1)",
    "flask (==2.2.5)",
    "fonttools (==4.49.0)",
    "holidays (==0.44)",
    "idna (==3.6)",
    "importlib-resources (==6.1.3)",
    "inflection (==0.5.1)",
    "itsdangerous (==2.1.2)",
    "jinja2 (==3.1.3)",
    "jsonschema (==4.21.1)",
    "jsonschema-specifications (==2023.12.1)",
    "kiwisolver (==1.4.5)",
    "markupsafe (==2.1.5)",
    "matplotlib (==3.8.3)",
    "numpy (==1.26.4)",
    "packaging (==24.0)",
    "pandas (==2.2.1)",
    "pillow (==10.2.0)",
    "prophet (==1.1.5)",
    "pyparsing (==3.1.2)",
    "python-dateutil (==2.9.0.post0)",
    "pytz (==2024.1)",
    "pyyaml (==6.0.1)",
    "referencing (==0.33.0)",
    "requests (==2.31.0)",
    "rpds-py (==0.18.0)",
    "six (==1.16.0)",
    "stanio (==0.3.0)",
    "tqdm (==4.66.2)",
    "tzdata (==2024.1)",
    "urllib3 (==2.2.1)",
    "werkzeug (==2.2.3)",
    "grpcio-tools (==1.62.1)",
    "grpcio (==1.62.1)",
    "google-api-python-client (==2.123.0)",
    "protoc-gen-openapiv2 (==0.0.1)",
    "questdb (==2.0.3)",
    "mysql-connector-python (==8.4.0)",
    "sqlalchemy (==2.0.30)",
    "pyarrow (==15.0.2)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "fire (==0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
bandit = "1.8.0"
pyright = "1.1.391"
pylint = "3.3.5"
isort = "5.13.2"
black = "24.10.0"
radon = "6.0.1"
codespell = "2.3.0"

